# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciar ONGs e pets",
        "title": "OngPet API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "description": "Faz login via Supabase e retorna tokens (access_token, etc.)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Credenciais",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/ongs": {
            "get": {
                "description": "Retorna todas as ONGs cadastradas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ONG"
                ],
                "summary": "Lista todas as ONGs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.OngListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cria uma ONG no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ONG"
                ],
                "summary": "Cria uma nova ONG",
                "parameters": [
                    {
                        "description": "Nova ONG",
                        "name": "ong",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreateOngInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Ong"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/ongs/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Atualiza os dados de uma ONG pelo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ONG"
                ],
                "summary": "Atualiza uma ONG existente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da ONG",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados para atualização",
                        "name": "ong",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreateOngInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "ong": {
                                    "$ref": "#/definitions/models.Ong"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove uma ONG pelo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ONG"
                ],
                "summary": "Remove uma ONG",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da ONG",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Ong": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "endereco": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "nome_responsavel": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "v1.CreateOngInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "a@b.com"
                },
                "endereco": {
                    "type": "string",
                    "example": "Rua A"
                },
                "nome": {
                    "type": "string",
                    "example": "Minha ONG"
                },
                "nome_responsavel": {
                    "type": "string",
                    "example": "Fulano"
                },
                "telefone": {
                    "type": "string",
                    "example": "123"
                }
            }
        },
...
        "v1.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "v1.OngListResponse": {
            "type": "object",
            "properties": {
                "dados": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ong"
                    }
                },
                "proxima_pagina": {
                    "type": "boolean"
                },
                "total_paginas": {
                    "type": "integer"
                },
                "total_registros": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}